refactor: Update hit effect types and improve component imports for combat system
