+++ src/components/combat/components/index.ts
@@ -1,14 +1,12 @@
 /**
- * @fileoverview Complete Combat Components exports for Black Trigram Korean martial arts game
+ * @fileoverview Complete Combat System exports for Black Trigram Korean martial arts game
+ * @description Provides comprehensive combat functionality including movement system,
+ * enhanced input handling, AI opponents, and traditional Korean martial arts mechanics
  */
 
 // Core combat screens and systems
-export { CombatScreen } from "../CombatScreen";
-export { default as CombatScreenDefault } from "../CombatScreen";
-
-// Combat engine and game logic
-export { GameEngine } from "../engine/GameEngine";
-export { default as GameEngineDefault } from "../engine/GameEngine";
+export { CombatScreen } from "./CombatScreen";
+export { default as CombatScreenDefault } from "./CombatScreen";
 
 // Combat components
 export { CombatArena } from "./CombatArena";
@@ -22,22 +20,22 @@ export { Player } from "./Player";
 export { PlayerVisuals } from "./PlayerVisuals";
 export { HitEffectsLayer } from "./HitEffectsLayer";
 
-// Background and environment
+// Fix: Import from correct locations
+export { GameEngine } from "../engine/GameEngine";
+export { default as GameEngineDefault } from "../engine/GameEngine";
+
 export { DojangBackground } from "../backgrounds/DojangBackground";
 export { default as DojangBackgroundDefault } from "../backgrounds/DojangBackground";
 
-// Layout components (fix import path)
-export { KoreanButton } from "../../ui/base/KoreanLayoutComponents";
-
-// Default exports for convenience
-export { default as CombatArenaDefault } from "./CombatArena";
-export { default as CombatHUDDefault } from "./CombatHUD";
-export { default as CombatControlsDefault } from "./CombatControls";
-export { default as CombatStatsDefault } from "./CombatStats";
-export { default as PlayerStatusPanelDefault } from "./PlayerStatusPanel";
-export { default as PlayerDefault } from "./Player";
-export { default as PlayerVisualsDefault } from "./PlayerVisuals";
-export { default as HitEffectsLayerDefault } from "./HitEffectsLayer";
+// Component index export - remove duplicates
+export * from "./CombatArena";
+export * from "./CombatHUD";
+export * from "./CombatControls";
+export * from "./CombatStats";
+export * from "./PlayerStatusPanel";
+export * from "./Player";
+export * from "./PlayerVisuals";
+export * from "./HitEffectsLayer";
 
 // Re-export types for combat system
 export type {
@@ -60,3 +58,5 @@ export type {
   HitEffect,
   DisplayHitEffect,
 } from "../../../types/";
+
+// ...existing combat features documentation...
